trying to figure out data grid

+++ Models/Domain/RecurringTransaction.cs
@@ -9,7 +9,7 @@ public abstract class RecurringTransaction
     public required string Name { get; set; }
     public required string Description { get; set; }
     public string? Notes { get; set; }
-    public required decimal Amount { get; set; }
+    public required double Amount { get; set; }
     public required DateTime StartDate { get; set; }
     public required DateTime EndDate { get; set; }
     public required RecurrancePattern Pattern { get; set; }
